fix(xai): align realtime Voice Agent defaults and API coverage
Default to grok-voice-latest with grok-transcribe, add force_message say(),
and handle live caption updates so the LiveKit plugin matches current xAI
Realtime behavior (and closes the Pipecat-era default-model gap).
