+++ livekit-plugins/livekit-plugins-xai/livekit/plugins/xai/types.py
@@ -58,9 +58,10 @@ GrokVoices = Literal[
 ]
 
 GrokRealtimeModels = Literal[
-    "grok-voice-fast-1.0",
-    "grok-voice-think-fast-1.0",
+    "grok-voice-latest",
     "grok-voice-think-fast-2.0",
+    "grok-voice-think-fast-1.0",
+    "grok-voice-fast-1.0",
 ]
 
 TTSLanguages = Literal[
